add proper err return and print

+++ cmd/destruct.go
@@ -17,7 +17,7 @@ package cmd
 
 import (
 	"errors"
-	"fmt"
+	"github.com/AlecAivazis/survey"
 	"os"
 	"path/filepath"
 
@@ -36,20 +36,48 @@ and usage of using your command. For example:
 Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
+	RunE: func(cmd *cobra.Command, args []string) error {
 		wd, err := os.Getwd()
 		if err != nil {
-			fmt.Fprintf(os.Stderr, err.Error())
-			return
+			return err
 		}
+
+		isSure := false
+		prompt := &survey.Confirm{
+			Message: "Are you sure you want to destruct? It will delete your .gud folder",
+		}
+		err = survey.AskOne(prompt, &isSure, icons)
+		if err != nil {
+			return err
+		}
+
+		if !isSure {
+			return nil
+		}
+
+		isSure2 := false
+		prompt = &survey.Confirm{
+			Message: "Can you please not to? It really wants to live...",
+		}
+		err = survey.AskOne(prompt, &isSure2, icons)
+		if err != nil {
+			return err
+		}
+
+		if !isSure2{
+			return nil
+		}
+
 		root, err := getRoot(wd)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, err.Error())
+			return err
 		}
 		_ = os.RemoveAll(root)
 		if restartF {
 			startCmd.Run(cmd, args)
 		}
+
+		return nil
 	},
 }
